requirements.txt
aiosqlite==0.20.0
alembic==1.8.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Eel==0.16.0
fastapi==0.110.0
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
pefile==2023.2.7
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.11
PyJWT==2.8.0
pyparsing==3.1.1
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32-ctypes==0.2.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.28.0
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1